Added more useful comments to the example project | extracted MyController to a new file
